Add support for custom allocators
Currently all allocations in GPOS (and hence GPORCA) go through
CMemoryPoolMalloc as the underlying memory allocator. This pull requests
introduces a new class CMemoryPoolCustomAlloc that holds two function
pointers to custom memory allocators that can be set up during during
the initialization of GPOS.

+++ libgpos/include/gpos/common/clibwrapper.h
@@ -1,6 +1,6 @@
 //---------------------------------------------------------------------------
 //	Greenplum Database
-//	Copyright (C) 2011 EMC Corp.
+//	Copyright (c) 2004-2015 Pivotal Software, Inc.
 //
 //	@filename:
 //	       	clibwrapper.h
@@ -126,7 +126,7 @@ namespace gpos
 		TIME *PtmLocalTimeR(const TIME_T *ptTime, TIME *ptmResult);
 
 		// allocate dynamic memory
-		void *PvMalloc(ULONG ulSize);
+		void *PvMalloc(SIZE_T ulSize);
 
 		// free dynamic memory
 		void Free(void *pvSrc);
